refactoring for LICs and bug discoveries #65
I refactored the headers and the test functions so everything is in the correct order and is consistent

When building and trying to run the compiled code however, I discovered a fragmentation error #63. The test will be commented out for now

I also discovered that LIC13 used Lic8() function, which when changed causes the unit test to fail #64.