Launch tests (#81)
* Fixed a bug in the code that made the calc_pum method to work
incorrectly. #74

* Positive test for the decide funtion #62

Test that checks if the program correctly sets launch to true
give valid inputs

* Another positive test case for the entire program. #62

Checks that launch is true if the LIC input is only not used

* Added a negative test case for the decide method. #62

* Added a statement of contributions to the readme file. #69

* Implemented a positive test for the program #62

This test checks that launch is set to true if all lics are
required to be true and they are.

* Added a negative test case for the entire program. #62

Checks that launch is false if all lics are required to be true,
and all are except for one.